// ams-server-Final_Source_BE/ams-portal-Final_Source_FE/src/app/enum/Trainee.enum.ts
export enum AllowanceGroup {
    DEV_N = 'Dev_N',
    DEV_J = 'Dev_J',
    DEV_S = 'Dev_S'
}

export enum TraineeStatus {
    WAITING_FOR_CLASS = 'WAITING_FOR_CLASS',
    ENROLLED = 'ENROLLED',
    DEFERRED = 'DEFERRED',
    WAITING_FOR_ALLOCATION = 'WAITING_FOR_ALLOCATION',
    ALLOCATED = 'ALLOCATED',
    DRAFT = 'DRAFT'
}

export enum TraineeStatusInClass {
    ACTIVE = 'ACTIVE',
    DROP_OUT = 'DROP_OUT',
    DEFERRED = 'DEFERRED'
}